# Video-Editor-Pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClipFlow Video Editor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/export:
    post:
      operationId: exportProject
      tags: [projects]
      summary: Export project as a video file using FFmpeg
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportProjectResponse"
        "400":
          description: No processable clips found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ExportProjectResponse:
      type: object
      properties:
        downloadUrl:
          type: string
          description: Signed URL to download the exported video (valid 24h)
        objectPath:
          type: string
          description: GCS object path for the exported video
        duration:
          type: number
          description: Duration of the exported video in seconds
      required:
        - downloadUrl
        - objectPath
        - duration
...